Made changes for authorization along with minor sql and logchanges
