+++ worker-service/sql/mysql-ers-worker.sql
@@ -23,6 +23,9 @@ alter table worker
 	add constraint UK_1pracyjlfmis4kre9cb740etf unique (username);
     
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('One', 'User', 'employee1', 'employee1', 'Data Coordiator', 10, '1/1/2019', '41 Randy Pass', 'Boston', 'Massachusetts', '02124', '611-772-4221', 'user1@admin.com');
+insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('One', 'Manager', 'manager1', 'manager1', 'Manager', 0, '1/1/2016', '43255 Pennman', 'St Louis', 'Missouri', '77544', '617-772-4221', 'manager1@yahoo.com');
+
+
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Ten', 'Gwen', 'benjamin722', 'benchee5252', 'Data Coordiator', 10, '1/1/2018', '42 Portland', 'Boston', 'Massachusetts', '22222', '612-772-4222', 'ben2@admin.ch');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Ten', 'Kevin', 'benjamin723', 'benchee5353', 'IT Tech', 10, '3/24/2017', '2314 Boulevard', 'Boston', 'Massachusetts', '33333', '613-772-4223', 'ben3@admin.ch');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Pope', 'Steven', 'benjamin724', 'benchee5454', 'IT Tech', 10, '5/2/2014', '66474 Randy Pass', 'Boston', 'Massachusetts', '12345', '614-772-4224', 'ben4@admin.ch');
@@ -31,7 +34,6 @@ insert into worker (last_name, first_name, username, pass_code, title, reports_t
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Dover', 'Ben', 'benjamin727', 'benchee5757', 'Accountant', 10, '1/13/2016', '41442 Lincoln Rd', 'Boston', 'Massachusetts', '11754', '617-772-4221', 'ben7@admin.ch');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Dover', 'Gwen', 'benjamin728', 'benchee5858', 'Accountant', 10, '1/24/2019', '256232 Randy Pass', 'Michigan City', 'Michigan', '5452`', '617-772-4221', 'ben8@admin.ch');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Tuff', 'Lucas', 'benjamin729', 'benchee5959', 'Accountant', 10, '1/13/2011', '32154 Yeti Ave', 'Ann Arbor', 'Michigan', '54786', '617-772-4221', 'ben9@admin.ch');
-insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('One', 'Manager', 'manager1', 'manager1', 'Manager', 0, '1/1/2016', '43255 Pennman', 'St Louis', 'Missouri', '77544', '617-772-4221', 'manager1@yahoo.com');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Hagger', 'Arny', 'ahagger0', 'ZIfvfpbYm', 'Developer',15 , '7/28/2019', '56248 Rowland Pass', 'Shawnee Mission', 'Kansas', '66220', '913-561-7825', 'ahagger0@digg.com');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('McKeand', 'Jennee', 'jmckeand1', 'OmN4mNbkk', 'Developer', 15, '5/6/2011', '95 Garrison Lane', 'Kansas City', 'Missouri', '64142', '816-995-5558', 'jmckeand1@ftc.gov');
 insert into worker (last_name, first_name, username, pass_code, title, reports_to, hire_date, street_address, city, state, zip, phone, email) values ('Burmaster', 'Brigitte', 'bburmaster2', 'RjHt1s', 'DBA', 15, '10/14/2019', '020 Muir Way', 'Boynton Beach', 'Florida', '33436', '561-135-8161', 'bburmaster2@devhub.com');
