+++ worker-service/src/main/java/com/ers/projects/controllers/LogInController.java
@@ -1,5 +1,6 @@
 package com.ers.projects.controllers;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,15 +19,21 @@ public class LogInController {
 	
 	@Autowired
 	private WorkerService ws;
-	
+	private Logger log = Logger.getLogger(LogInController.class);
 	
 	@PostMapping("/authenticate")
 	public Worker authorize(@RequestBody Worker w) {
+		System.out.println(w);
 		Worker worker = ws.getWorkerByUsername(w.getUsername());
-		if(worker.getUsername().equals(w.getUsername()) && worker.getPassCode().equals(w.getPassCode())) {
-			return worker;
+		if(worker != null) {
+			if(worker.getUsername().equals(w.getUsername()) && worker.getPassCode().equals(w.getPassCode())) {
+				log.info("Success: credentials for " + w.getUsername() + " matched.");
+				return worker;
+			}
+			else {
+				log.info("Fail: credentials for " + w.getUsername() + " did not match. Incorrect password");
+			}
 		}
 		return null;
 	}
-
 }
